fix: Correct return types and LIMIT queries in repositories
- RefreshTokenRepository: Change countActiveTokensByUser return type
  from int to long (COUNT returns Long in JPQL)
- RestaurantPayoutRepository: Change getAverageProcessingTimeHours
  return type from BigDecimal to Double (AVG returns Double)
- BackupLogRepository: Convert LIMIT queries to native SQL
  (findLastSuccessfulBackup, findLastBackup)

+++ src/main/java/com/fooddelivery/analytics/financial/repository/RestaurantPayoutRepository.java
@@ -83,8 +83,8 @@ public interface RestaurantPayoutRepository extends JpaRepository<RestaurantPayo
                    "FROM restaurant_payouts " +
                    "WHERE status = 'COMPLETED' " +
                    "AND period_start >= :startDate AND period_end <= :endDate", nativeQuery = true)
-    BigDecimal getAverageProcessingTimeHours(@Param("startDate") LocalDate startDate,
-                                             @Param("endDate") LocalDate endDate);
+    Double getAverageProcessingTimeHours(@Param("startDate") LocalDate startDate,
+                                         @Param("endDate") LocalDate endDate);
 
     /**
      * Get payouts for specific restaurants.
